pyproject.toml
[project]
name = "rag-patterns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.5",
    "faiss-cpu>=1.12.0",
    "graphviz>=0.21",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.30",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.6",
    "openai>=1.97.1",
    "pypdf>=5.8.0",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.0",
    "requests>=2.32.4",
    "rich>=14.1.0",
    "tavily-python>=0.7.11",
    "tqdm>=4.67.1",
    "urllib3>=2.5.0",
]
